Don't apply cache buster if no refresh value set
- Cachebuster only appended to url if a refresh value has been selected from settings
- Resolve jQuery conflict
- Test file renaming
- Appropriate tests
