+++ bower.json
@@ -40,6 +40,7 @@
   },
   "resolutions": {
     "angular": "~1.3.0",
-    "angular-mocks": "~1.3.0"
+    "angular-mocks": "~1.3.0",
+    "jquery": "~2.1.4"
   }
 }
