+++ test/data/refresh.js
@@ -1,0 +1,5 @@
+(function(window) {
+  "use strict";
+
+  window.gadget.settings.additionalParams.refresh = 1;
+})(window);
